### 2021-03-05 #432 @mccallc
submission: awardFindR
Submitting Author: Michael McCall (@mccallc)
Other Authors: Sebastian Karcher (@adam3smith)
Repository:  https://github.com/PESData/awardFindR
Version submitted: 0.1.0
Version accepted: TBD

---

```
Package: awardFindR
Title: QDR awardFindR
Version: 0.1.0
Authors@R: c(person("Michael", "McCall", email = "mimccall@syr.edu",
                  role = c("aut", "cre"), comment = c(ORCID="0000-0002-4668-4212")),
                  person("Sebastian", "Karcher", email = "karcher@u.northwestern.edu",
                  role = c("ctb"), comment = c(ORCID = "0000-0001-8249-7388")),
                  person("Qualitative Data", "Repository", email = "qdr@syr.edu",
                  role = c("cph")),
                  person("Sloan", "Foundation", email = "technology@sloan.org",
                  role = c("fnd")))
Author: Michael C. McCall
Maintainer: Michael McCall <mimccall@syr.edu>
Description: Queries a number of scientific awards databases.
    Collects relevant results based on keyword and date parameters,
    returns list of projects that fit those criteria as a data frame.
    Sources include: Arnold Ventures, Carnegie Corp, Federal RePORTER,
    Gates Foundation, MacArthur Foundation, Mellon Foundation, NEH, NIH,
    NSF, Open Philanthropy, Open Society Foundations, Rockefeller Foundation,
    Russell Sage Foundation, Robert Wood Johnson Foundation,
    Sloan Foundation, Social Science Research Council, John Templeton Foundation,
    USASpending.gov
    rvest,
    xml2,
    httr
    vcr,
    covr
```

Acts as an interface to web-based APIs and search engines.

Anyone interested in trends in academic funding or researchers looking for a suitable funder based on previously funded research. The package can be used for scientometrics more generally, or more narrowly field-specific lines of inquiry.

The target audience is primarily US-based researchers, and the sources of governmental research funding are all covered to reasonable expectations through the combination of NSF, NEH, NIH, Federal RePORTER and USAspending. Private funders are more difficult to evaluate, but the sources included are the most prominent in the social sciences.

Nothing that is functionally similar. Some source-specific packages exist, but none that aggregate across sources.

[nihexporter](https://github.com/jayhesselberth/nihexporter) provides search functionality for the NIH grant database, but is limited to an existing scrape of NIH data from 2016. [awardsBot](https://github.com/NCEAS/awards-bot) dynamically scrapes NSF awards, but lacks search functionality and is primarily designed for tracking and emailing PIs regarding their obligations. [fedreporter](https://github.com/muschellij2/fedreporter) is not actively maintained and only offers search functionality for a antiquated and limited version of the Federal Reporter API we implement here.

This package offers a unified, simple search functionality across these federal grant databases, and can include many private funders as well that are not covered by any existing packages.

Yes
