### 2021-03-10 #432 @melvidoni
Hello @mccallc, thanks for submitting. We were discussing this submission with the Associate Editors, but we need more information.

In particular, we want to know "how much this package solves the problem". Say, does it cover all possible sources? The top-10 sources? The most prestigious sources? Also, it would be nice to know not just that other packages cover some sources, but more precisely what proportion of the covered sources are already provided by other packages. If all sources are already considered, maybe the benefit of this package is that it might provide a consistent interface across sources?

Please, edit your answer on the submission post, and let me know once it is done.
