### 2021-03-24 #432 @annakrystalli
- [x] **Automated tests:** Package has a testing suite and is tested via a CI service.

---

Thanks for submitting this super interesting package! I've now completed initial editor checks and there are a few issues to sort out before we proceed.

##### goodpractice

There are a number of issues returned from running `goodpractice::gp()`

**It is good practice to**

- [ ] .

    ```
    ✖ avoid long code lines, it is bad for readability.
    Also, many people prefer editor windows that are about 80
    characters wide. Try make your lines shorter than 80
    characters

    R/arnold.R:10:1
    R/arnold.R:20:1
    R/carnegie.R:34:1
    R/carnegie.R:35:1
    R/fedreporter.R:5:1
    ```

- [ ] .

    ```
    ✖ avoid sapply(), it is not type safe. It might
    return a vector, or a list, depending on the input data.
    Consider using vapply() instead.

    R/carnegie.R:25:13
    R/main.R:89:26
    R/main.R:90:17
    R/rockefeller.R:21:18
    ```

- [ ] .

    ```r
    ✖ fix this R CMD check NOTE: Malformed Description
    field: should contain one or more complete sentences.
    ```

- [ ] .

    ```
    ✖ fix this R CMD check NOTE: Namespace in Imports
    field not imported from: ‘readr’ All declared Imports should
    be used.
    ```

- [ ] .

    ```
    ✖ fix this R CMD check ERROR: Running examples in
    ‘awardFindR-Ex.R’ failed The error most likely occurred in: >
    ### Name: gates_get > ### Title: Query awards from the Bill &
    Melinda Gates Foundation > ### Aliases: gates_get > > ### **
    Examples > > gates <- gates_get("qualitative", "2018-01-01",
    "2020-01-01") POST
    https://www.gatesfoundation.org/services/gfo/search.ashx ...
    Not Found (HTTP 404). Warning in request(url, "post",
    payload) : Not Found (HTTP 404). Error: $ operator is invalid
    for atomic vectors Execution halted
    ```

- [ ] .

    ```r
    ✖ checking tests ... Running ‘testthat.R’ ERROR
    Running the tests in ‘tests/testthat.R’ failed. Last 13 lines
    of output: 11.  └─awardFindR:::FUN(X[[i]], ...) 12.
    └─awardFindR:::request(paste0(query_url, "&offset=", offset),
    "get") 13.  └─httr::GET(url) 14.
    └─httr:::request_perform(req, hu$handle$handle) 15.
    └─httr:::perform_callback("request", req = req) 16.
    └─webmockr:::callback(...) 17.
    └─webmockr::HttrAdapter$new()$handle_request(req) 18.
    └─private$request_handler(req)$handle() 19.
    └─eval(parse(text = req_type_fun))(self$request) 20.
    └─err$run() 21.  └─self$construct_message() [ FAIL 1 | WARN 0
    | SKIP 0 | PASS 14 ] Error: Test failures Execution halted
    ```

- [ ] .

    ```r
    ✖ avoid 'T' and 'F', as they are just variables which
    are set to the logicals 'TRUE' and 'FALSE' by default, but
    are not reserved words and hence can be overwritten by the
    user.  Hence, one should always use 'TRUE' and 'FALSE' for
    the logicals.

    R/arnold.R:NA:NA
    R/arnold.R:NA:NA
    R/arnold.R:NA:NA
    R/carnegie.R:NA:NA
    R/carnegie.R:NA:NA
    ```

##### Tests

A bit more detail on the single failing test when running `devtools::test()` (also turning up in `devtools::check()`)

```
Error (test-full.R:2:3): Search that should return all empty

================================================================================
An HTTP request has been made that vcr does not know how to handle:
GET https://api.nsf.gov/services/v1/awards.json?keyword="foobar"&cfdaNumber=47.041,47.050,47.049,47.078,47.083,47.079,47.070,47.074,47.076,47.075&printFields=id,date,startDate,expDate,title,awardeeName,piFirstName,piLastName,piEmail,cfdaNumber,fundsObligatedAmt,fundProgramName&dateStart=01/01/2019&dateEnd=03/22/2021&offset=1
vcr is currently using the following cassette:
  - ../fixtures/empty.yml
    - record_mode: once
    - match_requests_on: method, uri
Run `vcr::vcr_last_error()` for more verbose errors
If you're not sure what to do, open an issue https://github.com/ropensci/vcr/issues
& see https://books.ropensci.org/http-testing
================================================================================

  1. base::suppressMessages(...) test-full.R:2:2
  5. awardFindR::awardFindR("foobar") test-full.R:3:4
  9. awardFindR::.nsf_standardize(keywords, from_date, to_date)
 10. base::lapply(keywords, nsf_get, from_date, to_date) /Users/Anna/Documents/workflows/rOpenSci/editorials/awardFindR/R/nsf.R:49:2
 11. awardFindR:::FUN(X[[i]], ...)
 12. awardFindR::request(paste0(query_url, "&offset=", offset), "get") /Users/Anna/Documents/workflows/rOpenSci/editorials/awardFindR/R/nsf.R:24:2
 13. httr::GET(url) /Users/Anna/Documents/workflows/rOpenSci/editorials/awardFindR/R/utils.R:30:4
 14. httr:::request_perform(req, hu$handle$handle)
 15. httr:::perform_callback("request", req = req)
 16. webmockr:::callback(...)
 17. webmockr::HttrAdapter$new()$handle_request(req)
 18. private$request_handler(req)$handle()
 19. eval(parse(text = req_type_fun))(self$request)
 20. err$run()
 21. self$construct_message()
──────────────────────────────────────────────────────────────────────────
✓ |   1       | http [0.3 s]
✓ |   1       | neh [1.7 s]
✓ |   1       | ophil [0.2 s]
✓ |   1       | osociety [0.4 s]
✓ |   2       | rockefeller [0.4 s]
✓ |   1       | ssrc [0.9 s]
✓ |   1       | templeton [3.7 s]
✓ |   1       | usaspend [0.2 s]

══ Results ═══════════════════════════════════════════════════════════════
```
Test failure means I can't run `covr::package_coverage()` to get a breakdown of test coverage.

### Spellcheck

Only a single typo thrown up by `devtools::spell_check()`

- [ ] .

    ```
    Philanthrophy   README.md:23
    ```

### Documentation question

In your docs, you only demonstrate the use of the `awardFindR()` function, yet in the reference section, the list of exported functions is much longer, including finctions like `request()` that seems to be more of a back-end function. It's a good time to review exported functions to make sure the reviewers can focus on functions that are only relevant to the end user. Also, for the functions you do decide to export, while not necessary to introduce every single function in the docs (that's what examples and the reference section are for after all!), it would be good to mention to the user in the docs that other exported functions exist with an example of how they could typically use one of them.

---

@mccallc, could you also please add the rOpenSci under review badge to your README?
