+++ tests/testthat/test-http.R
@@ -1,7 +1,8 @@
 suppressMessages(vcr::use_cassette("http", {
 
 test_that("HTTP GET works to fetch html", {
-  expect_match(attributes(suppressMessages(request("http://r-project.org", "get")))$class[1],
+  expect_match(attributes(suppressMessages(
+    request("http://r-project.org", "get")))$class[1],
                "xml_document")
 })
 
