+++ R/nih.R
@@ -33,13 +33,16 @@ nih_get <- function(keyword, from_date, to_date) {
 
   # change NULL values to NA
   df <- lapply(response$results, lapply, function(x)ifelse(is.null(x), NA, x))
-  df <- do.call(rbind.data.frame, df)
+  df <- rbind.match.columns(df)
 
   # What is our current max?
   place <- response$meta$offset + response$meta$limit
   # Do we need to loop with different offsets?
   while (place < response$meta$total) {
     payload$offset <- place
+
+    Sys.sleep(3) # Be NICE to the API!
+
     response <- request(url, "post", payload) # Query again
     place <- response$meta$offset + response$meta$limit # New position
 
@@ -47,7 +50,7 @@ nih_get <- function(keyword, from_date, to_date) {
     temp <- lapply(response$results, lapply,
                    function(x)ifelse(is.null(x), NA, x))
 
-    temp <- do.call(rbind.data.frame, temp)
+    temp <- rbind.match.columns(temp)
     df <- rbind.data.frame(df, temp)
   }
 
@@ -56,6 +59,9 @@ nih_get <- function(keyword, from_date, to_date) {
   # Remove factors
   df[] <- lapply(df, as.character)
   df$keyword <- keyword
+
+  # NIH keywords are autogenerated? check for false positives in the abstract
+  df <- df[grepl(keyword, df$abstract_text, ignore.case = TRUE), ]
 
   df
 }
