+++ tests/fixtures/macarthur.yml
@@ -1,0 +1,715 @@
+http_interactions:
+- request:
+    method: get
+    uri: https://searchg2.crownpeak.net/live-macfound-rt/select?q=qualitative&wt=xml&start=0&rows=100&echoParams=explicit&fl=*&defType=edismax&fq=custom_s_template:%22grant%20detail%22&sort=score%20desc&qf=custom_t_title%20custom_t_description%20custom_t_name&indent=true&json.wrf=searchg2_5445608496089546
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      Accept: application/json, text/xml, application/xml, */*
+  response:
+    status:
+      status_code: 200
+      category: Success
+      reason: OK
+      message: 'Success: (200) OK'
+    headers:
+      access-control-allow-origin: '*'
+      content-type: application/xml; charset=utf-8
+      date: Fri, 26 Mar 2021 17:36:05 GMT
+      server: Apache/2.4.3 (Unix) OpenSSL/1.0.2g
+      transfer-encoding: chunked
+      connection: keep-alive
+    body:
+      encoding: UTF-8
+      file: no
+      string: |
+        <?xml version="1.0" encoding="UTF-8"?>
+        <response>
+
+        <lst name="responseHeader">
+          <bool name="zkConnected">true</bool>
+          <int name="status">0</int>
+          <int name="QTime">0</int>
+          <lst name="params">
+            <str name="q">qualitative</str>
+            <str name="defType">edismax</str>
+            <str name="json.wrf">searchg2_5445608496089546</str>
+            <str name="echoParams">explicit</str>
+            <str name="qf">custom_t_title custom_t_description custom_t_name</str>
+            <str name="indent">true</str>
+            <str name="fl">*</str>
+            <str name="start">0</str>
+            <str name="fq">custom_s_template:"grant detail"</str>
+            <str name="sort">score desc</str>
+            <str name="rows">100</str>
+            <str name="wt">xml</str>
+          </lst>
+        </lst>
+        <result name="response" numFound="14" start="0">
+          <doc>
+            <str name="id">87682</str>
+            <str name="url">87682</str>
+            <str name="custom_s_template">grant detail</str>
+            <str name="metadata_title">To support the project Blending Quantitative and Qualitative Data in the Study of Well-Being.</str>
+            <str name="metadata_description">The MacArthur Foundation supports creative people and effective institutions committed to building a more just, verdant, and peaceful world. In addition to selecting the MacArthur Fellows, we work to defend human rights, advance global conservation and security, make cities better places, and understand how technology is affecting children and society.</str>
+            <str name="metadata_keywords"> foundation, philanthropy, grant, grantmaking</str>
+            <str name="custom_s_foundation_id">56995-0</str>
+            <str name="custom_s_org_id">29416</str>
+            <str name="custom_s_parent_grantee_id"/>
+            <str name="custom_s_grantee_location">Madison, Wisconsin</str>
+            <str name="custom_s_grant_approved_date">1999-01-19T00:00:00</str>
+            <str name="custom_s_name">University of Wisconsin - Madison, Institute on Aging</str>
+            <str name="custom_t_name">University of Wisconsin - Madison, Institute on Aging</str>
+            <str name="custom_s_city">Madison</str>
+            <str name="custom_s_state">WI</str>
+            <str name="custom_s_state_display">Wisconsin</str>
+            <str name="custom_s_country">US</str>
+            <str name="custom_s_country_display">United States</str>
+            <str name="custom_s_amount">25000.0</str>
+            <str name="custom_s_amount_display">$25,000</str>
+            <str name="custom_s_title">To support the project Blending Quantitative and Qualitative Data in the Study of Well-Being.</str>
[…]
