### 2021-03-31 29a93b3
include abstract text variable
