+++ R/nih.R
@@ -77,6 +77,7 @@ nih_get <- function(keyword, from_date, to_date) {
     institution=org_name, pi=contact_pi_name, year=fiscal_year,
     start=project_start_date, end=project_end_date,
     program=agency_code, amount=as.integer(award_amount), id=project_num,
-    title=project_title, keyword, source="NIH", stringsAsFactors = FALSE
+    title=project_title, abstract=abstract_text,
+    keyword, source="NIH", stringsAsFactors = FALSE
   ))
 }
