### 2021-03-31 8976e35
remove internal function documentation
