+++ man/request.Rd
@@ -1,21 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
-\name{request}
-\alias{request}
-\title{Make an HTTP request with httr}
-\usage{
-request(url, method, payload = NULL)
-}
-\arguments{
-\item{url}{URL}
-
-\item{method}{"get" or "post"}
-
-\item{payload}{list object to convert to json and send, only if method="post"}
-}
-\value{
-Content of the HTTP response, as returned by httr::content()
-}
-\description{
-Make an HTTP request with httr
-}
