+++ NAMESPACE
@@ -20,4 +20,5 @@ export(ssrc_get)
 export(templeton_df)
 export(templeton_get)
 export(usaspend_get)
+importFrom(readr,read_csv)
 importFrom(rvest,"%>%")
