### 2021-04-02 #432 @mccallc
Hello @annakrystalli,

I believe I have addressed each of your points regarding good practice, tests, spellcheck, documentation, and adding the rOpenSci badge.

``goodpractice::gp()`` still returns a few lines missing tests, though these are mainly for fringe cases of error handling that can be reasonably expected to work. Over 99% of code is covered. If a full 100% is needed, please let me know.
