+++ tests/fixtures/gates.yml
@@ -1,61 +1,0 @@
-http_interactions:
-- request:
-    method: get
-    uri: https://www.gatesfoundation.org/api/grantssearch?date&displayedTaxonomy&listingId=d2a41504-f557-4f1e-88d6-ea109d344feb&loadAllPages=true&page=1&pageId=31242fca-dcf8-466a-a296-d6411f85b0a5&perPage=999&q=qualitative&sc_site=gfo&showContentTypes=false&showDates=false&showImages&showSummaries=false&sortBy=date-desc&sortOrder=desc&yearAwardedEnd=2019&yearAwardedStart=2018
-    body:
-      encoding: ''
-      string: ''
-    headers:
-      Accept: application/json, text/xml, application/xml, */*
-  response:
-    status:
-      status_code: 200
-      category: Success
-      reason: OK
-      message: 'Success: (200) OK'
-    headers:
-      cache-control: no-cache
-      connection: keep-alive
-      content-encoding: gzip
-      content-length: '1160'
-      content-type: application/json; charset=utf-8
-      date: Fri, 26 Mar 2021 17:35:31 GMT
-      expires: Fri, 26 Mar 2021 17:35:31 GMT
-      pragma: no-cache
-      server: Microsoft-IIS/10.0
-      server-timing:
-      - cdn-cache; desc=MISS
-      - edge; dur=11
-      - origin; dur=113
-      vary: Accept-Encoding
-    body:
-      encoding: UTF-8
-      file: no
-      string: '{"request":{"pageId":"31242fca-dcf8-466a-a296-d6411f85b0a5","page":1,"perPage":999,"sortBy":"date-desc","sortOrder":"desc","queryParameters":{"loadAllPages":"true","q":"qualitative","yearAwardedEnd":"2019","yearAwardedStart":"2018"}},"totalResults":3,"results":[{"id":"{558B7F10-C09B-46D1-BAC3-BD5F70FFBC0B}","grantee":"Olson
-        Zaltman","division":"U.S. Program","divisionExtras":0,"mainCat":null,"grantTopic":"U.S.
-        Economic Mobility & Opportunity","grantTopicExtras":0,"yearAwarded":"2019","grantStatus":"Closed","grantDuration":"5","granteeLocation":{"city":"Pittsburgh","state":"Pennsylvania","country":"United
-        States"},"regionServed":"GLOBAL","regionServedExtras":1,"awardedAmount":"$231,444","url":"/about/committed-grants/2019/11/inv003474"},{"id":"{E4047BB4-D2EA-495E-BFC4-689FEABDA9FE}","grantee":"Action
-        Contre La Faim","division":"Global Development","divisionExtras":0,"mainCat":null,"grantTopic":"Nutrition","grantTopicExtras":0,"yearAwarded":"2019","grantStatus":"Active","grantDuration":"24","granteeLocation":{"city":"Paris","state":"","country":"France"},"regionServed":"GLOBAL","regionServedExtras":2,"awardedAmount":"$1,500,000","url":"/about/committed-grants/2019/08/opp1208004"},{"id":"{1819B763-EC95-4767-A37C-B20FFAD1004B}","grantee":"Princeton
-        University","division":"U.S. Program","divisionExtras":0,"mainCat":null,"grantTopic":"U.S.
-        Economic Mobility & Opportunity","grantTopicExtras":0,"yearAwarded":"2018","grantStatus":"Closed","grantDuration":"19","granteeLocation":{"city":"Princeton","state":"New
-        Jersey","country":"United States"},"regionServed":"GLOBAL","regionServedExtras":1,"awardedAmount":"$91,845","url":"/about/committed-grants/2018/11/opp1203985"}],"facets":[{"dropdownSelectLabel":"Select
[…]
