### 2021-05-29 #432 @karawoo
- [ ] **Examples** (that run successfully locally) for all exported functions -- one example threw an error (see details below)

---

This is a really nice and well put together package. It was very easy to get started with it and the functions are quite intuitive. I thought the contributing guide was also particularly well written.

While testing the examples I found that one example threw an error, and a few others didn't error but returned different output than I expected:

``` r
library("awardFindR")

## Throws an error
specific <- awardFindR(keywords=c("ethnography", "case studies"))
#> ... [removed long curl output here] ...
#> Error in data.frame(grantee = info[2], date, amount, program = info[4],  :
#>   arguments imply differing number of rows: 1, 0

## Returns NULL
(arnold <- arnold_get("qualitative data", 2016, 2017))
#> POST https://pyj9b8sltv-dsn.algolia.net/1/indexes/*/queries?x-algolia-agent=Algolia%20for%20JavaScript%20(4.5.1)%3B%20Browser%20(lite)%3B%20instantsearch.js%20(4.8.3)%3B%20JS%20Helper%20(3.2.2)&x-algolia-api-key=bec9ead5977b11ae383f4df272c2f106&x-algolia-application-id=PYJ9B8SLTV ... OK (HTTP 200).
#> NULL

## Succeeds with a warning
sloan <- sloan_get(c("qualitative data", "case studies"), 2018, 2020)
#> GET https://sloan.org/grants-database?dynamic=1&order_by=approved_at&order_by_direction=desc&limit=3000 ... OK (HTTP 200).
#> Warning in grepl(keyword, descriptions, ignore.case = TRUE): argument 'pattern'
#> has length > 1 and only the first element will be used

## Returns a list instead of a data frame
nsf <- nsf_get("ethnography", "2020-01-01", "2020-02-01")
#> GET https://api.nsf.gov/services/v1/awards.json?keyword="ethnography"&printFields=id,date,startDate,expDate,title,awardeeName,piFirstName,piLastName,piEmail,cfdaNumber,fundsObligatedAmt,fundProgramName,abstractText,awardeeCounty&dateStart=01/01/2020&dateEnd=02/01/2020&offset=1 ... OK (HTTP 200).
class(nsf)
#> [1] "list"
```

<sup>Created on 2021-05-28 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9001)</sup>

I noticed that the `awardFindR()` sets the `stringsAsFactors` option to `FALSE`, and this will persist to the rest of the user's R session. While `FALSE` is now the default from R 4.0.0, I think it would be more polite to set the option within the function and restore the user's original settings when the function exits:

```r
opt <- options(stringsAsFactors = FALSE)
on.exit(options(opt))
```

In general the functions in this package are quite consistent, however some functions (`neh_get()` for example) provide a message when no results are found, whereas others do not. Personally I find the messages helpful (moreso than the HTTP requests that are printed).

#### Other minor comments

A few thoughts/ideas that I'll leave up to the package author whether they want to incorporate:

* Speaking of the HTTP request messages, I think these are probably sometimes useful, but there might also be times when they might not be wanted. While a user could wrap their function call in `suppressMessage()`, another option might be to add a `verbose` argument to the functions that shows all the requests if `TRUE` and not if `FALSE`. This is just an idea though -- ultimately that's a design decision for the author.

* In `gates_get()` there are arguments `from_year` and `to_year` defined as "Year integer to begin search", but the examples in the documentation show character strings in YYYY-MM-DD format. It looks like `gates_get()` can take full dates, so might want to use `from_date`/`to_date` and could use `@inheritParams` to inherit documentation from another function with the same arguments.

* While testing out `awardFindR()` I accidentally included a typo in my sources and entered them like this `c("nsf, fedreporter")` (all as one string). This was obviously user error, but a quick and easy way to return a more helpful error message could be to add

  ```r
  sources <- match.arg(sources, several.ok = TRUE)
  ```

  near the top of the function. Then my typo would have created the following error message:

  ```
  Error in match.arg(sources, several.ok = TRUE) :
    'arg' should be one of "arnold", "carnegie", "fedreporter", "gates", "macarthur", "mellon", "neh", "nih", "nsf", "ophil", "osociety", "rockefeller", "rsf", "rwjf", "sloan", "ssrc", "templeton", "usaspend"
  ```

  instead of

  ```
  Error in awardFindR("baikal", sources = c("nsf, fedreporter"), from_date = "2010-01-01",  :
    exists(paste0(".", source, "_standardize")) is not TRUE
  ```

  Up to the author if they want to include this, just a suggestion to help out clumsy-typing users like me :)
