### 2021-05-31 PESData/awardFindR#24 @adam3smith
RopenSci Kara Woo review
https://github.com/ropensci/software-review/issues/432#issuecomment-850915687

- [x] Fix all noted errors/unexpected output in examples
- [x] Don't force `stringAsFactor` on user session
- [x] consistent messages across all functions: include message when no results as in NEH
- [x] don't show HTTP message unless a new `verbose` argument is set to `TRUE`
- [x] Fix gates documentation and allow full dates in search
- [x] Better error message for invalid arguments/organizations.
