### 2021-06-01 #432 @zambujo
- [x] A statement of need clearly stating problems the software is designed to solve and its target audience in README
- [x] Installation instructions: for the development version of package and any non-standard dependencies in README
- [x] Vignette(s) demonstrating major functionality that runs successfully locally
- [x] Function Documentation: for all exported functions
- [x] Examples (that run successfully locally) for all exported functions
- [x] Community guidelines including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogenerated via Authors@R).

- [x] Installation: Installation succeeds as documented.
- [x] Functionality: Any functional claims of the software been confirmed.
- [x] Performance: Any performance claims of the software been confirmed.
- [x] Automated tests: Unit tests cover essential functions of the package
- [x] Packaging guidelines: The package conforms to the rOpenSci packaging guidelines

---

`{awardFindR}` is a well-organised, well-tested package providing a common interface to access scattered data on grants and awards over a series of APIs and websites.

I could not find any issues other than those pointed out in the previous reviews.  I will nevertheless add a few minor points, mostly about design and styling, whose implementation is not required and left to the consideration of the authors.

- `awardFindR` is both the name of the package and the name of the main function.  While this is not an issue, I think it is nice when `?awardFindR` returns the help page for the package as a whole, in which case I would consider renaming the main function'

- given the package structure, it could make sense to use Roxygen's inheritance functionality `@inheritParams` to document parameters only once in a common/generic function (in `main.R`, for instance);

- I have the impression that it is common to name functions with the verb prefixing the subject, as it follows a logic of inheritance (e.g. `print()`, `print.lm()`); so I was wondering whether it would not be more natural to reverse the function names (e.g. `neh_get()` would become `get_neh()`);

- I am not used to see the installation instructions towards the end to the `README.md`.  I am used to find them at the top of the file, right after a very brief description of the package.

- I could not find the use for `read_csv()` in `utils.R` (nor, accordingly, the need to import `readr`, in the `DESCRIPTION`);
