### 2021-06-01 PESData/awardFindR#25 @adam3smith
zambujo's RopenSci Review
https://github.com/ropensci/software-review/issues/432#issuecomment-851939864

Definitely do:

- [x] Use Roxygen inherit functionality for documentation
- [x] Move installation instructions toward top of readme (I'd say right after the initial paragraph) + add heading for supported sites and maybe move those down?
- [x] Check on use of `read_csv` and `readr`

Discuss

- [x] Change main function name to not clash w package name (I'd be inclined to to this)
- [x] Change verb/subject order of function from `neh_get` to `get_neh`. Check whether this is indeed more common. Practically, upside for using `neh` first is quicker auto-complete. Upside of using `get_` first is that you see all `get_` functions as part of autocomplete. I could go either way here. Can also ask on Twitter.
