### 2021-06-02 5e16df8
don't set stringsAsFactors in session
