### 2021-06-02 2decd62
implementing a verbose argument in all functions
