+++ tests/testthat/test-ssrc.R
@@ -1,6 +1,7 @@
 test_that("Expected results", {
   suppressMessages(vcr::use_cassette("ssrc", {
-    ssrc <- .ssrc_standardize("qualitative", "2012-01-01", "2020-01-01")
+    ssrc <- .ssrc_standardize("qualitative", "2012-01-01", "2020-01-01",
+                              FALSE)
   }))
   expect_equal(nrow(ssrc), 48)
 })

[…]
