### 2021-06-02 11902a2
test with verbose messages
