### 2021-06-02 ae9bae6
proper gates argument documentation
