### 2021-06-04 a5e7bf0
use roxygen2 inheritParams for consistency
