+++ R/ssrc.R
@@ -34,10 +34,7 @@
 }
 
 #' Search SSRC fellowships and grants by keyword and date
-#' @param keyword Keyword to query, single string
-#' @param from_year Year to begin search, integer
-#' @param to_year Year to end search, integer
-#' @param verbose enable verbose HTTP messages. TRUE/FALSE, default: false
+#' @inheritParams fedreporter_get
 #' @return a data.frame
 #' @export
 #' @examples ssrc <- ssrc_get("qualitative", 2015, 2016)
