+++ R/osociety.R
@@ -1,8 +1,5 @@
 #' Query and scrape Open Society foundation awards
-#' @param keyword Keyword, single string
-#' @param from_year Year to begin search, integer
-#' @param to_year Year to end search, integer
-#' @param verbose enable verbose HTTP messages. TRUE/FALSE, default: false
+#' @inheritParams fedreporter_get
 #' @return a data.frame
 #' @export
 #' @examples osociety <- osociety_get("qualitative", 2016, 2019)
