+++ R/nsf.R
@@ -54,6 +54,7 @@ nsf_get <- function(keyword, from_date, to_date, verbose=FALSE, cfda=NULL) {
   raw <- lapply(keywords, nsf_get, from_date, to_date, verbose)
   raw <- do.call(rbind.data.frame, raw)
   if (nrow(raw)==0) {
+    message("No results from NSF")
     return(NULL)
   }
 
