+++ R/ophil.R
@@ -60,6 +60,7 @@ ophil_get <- function(keyword, from_year, to_year, verbose=FALSE) {
                 verbose)
   raw <- do.call(rbind.data.frame, raw)
   if (nrow(raw)==0) {
+    message("No results from Open Philanthropy")
     return(NULL)
   }
 
