+++ R/rockefeller.R
@@ -30,6 +30,7 @@ rockefeller_get <- function(keyword, from_date, to_date, verbose) {
   raw <- lapply(keywords, rockefeller_get, from_date, to_date, verbose)
   raw <- do.call(rbind.data.frame, raw)
   if (nrow(raw)==0) {
+    message("No results from Rockefeller")
     return(NULL)
   }
 
