+++ R/gates.R
@@ -57,6 +57,7 @@ gates_get <- function(keyword, from_year, to_year, verbose=FALSE) {
                  verbose)
    raw <- do.call(rbind.data.frame, raw)
    if (nrow(raw)==0) {
+      message("No results from Gates")
       return(NULL)
    }
 
