+++ R/carnegie.R
@@ -77,6 +77,7 @@ carnegie_get <- function(keyword, from_year, to_year, verbose=FALSE) {
                 verbose)
   raw <- do.call(rbind.data.frame, raw)
   if (nrow(raw)==0) {
+    message("No results from Carnegie")
     return(NULL)
   }
 
