+++ R/mellon.R
@@ -53,6 +53,7 @@ mellon_get <- function(keyword, from_year, to_year, verbose=FALSE) {
                 verbose)
   raw <- do.call(rbind.data.frame, raw)
   if (nrow(raw)==0) {
+    message("No results from Mellon")
     return(NULL)
   }
 
