+++ R/ssrc.R
@@ -83,6 +83,7 @@ ssrc_get <- function(keyword, from_year, to_year, verbose=FALSE) {
                 verbose)
   raw <- do.call(rbind.data.frame, raw)
   if (nrow(raw)==0) {
+    message("No results from SSRC")
     return(NULL)
   }
 
