+++ R/osociety.R
@@ -61,6 +61,7 @@ osociety_get <- function(keyword, from_year, to_year, verbose=FALSE) {
                 verbose)
   raw <- do.call(rbind.data.frame, raw)
   if (nrow(raw)==0) {
+    message("No results from Open Society")
     return(NULL)
   }
 
