### 2021-06-04 3ef0d31
fix bugs in examples
