+++ man/sloan_get.Rd
@@ -24,6 +24,6 @@ Search for a set of keywords in the Sloan grants database.
 \examples{
 
 \dontrun{
-sloan <- sloan_get(c("qualitative data", "case studies"), 2018, 2020)
+sloan <- sloan_get("case studies", 2018, 2020)
 }
 }
