+++ R/utils.R
@@ -1,5 +1,3 @@
-#' @importFrom readr read_csv
-
 # Internal, return basic text hash
 # *NOT* cryptographic!! this is a toy hasher!
 .text_hash <- function(string) {
