+++ man/award_search.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/main.R
-\name{awardFindR}
-\alias{awardFindR}
+\name{award_search}
+\alias{award_search}
 \title{awardFindR grant search}
 \usage{
-awardFindR(
+award_search(
   keywords,
   sources = c("arnold", "carnegie", "fedreporter", "gates", "macarthur", "mellon",
     "neh", "nih", "nsf", "ophil", "osociety", "rockefeller", "rsf", "rwjf", "sloan",
@@ -34,16 +34,16 @@ These queries can be limited by keyword, source and date terms.
 }
 \examples{
 # Results for "ethnography" from NSF between 1/1 and 2/1 2020
-\dontrun{awards <- awardFindR("ethnography", "nsf",
+\dontrun{awards <- award_search("ethnography", "nsf",
 "2020-01-01", "2020-02-01")}
 
 # More intensive queries
 \dontrun{
 # Specific keywords, all sources:
-specific <- awardFindR(keywords=c("ethnography", "case studies"))
+specific <- award_search(keywords=c("ethnography", "case studies"))
 
 # Specific keyword, all sources, specific date range:
-five_years <- awardFindR(keywords="qualitative",
+five_years <- award_search(keywords="qualitative",
 from_date="2015-01-01", to_date="2020-01-01")
 }
 }
