+++ NAMESPACE
@@ -1,7 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(arnold_get)
-export(awardFindR)
+export(award_search)
 export(carnegie_get)
 export(fedreporter_get)
 export(gates_get)
@@ -20,5 +20,4 @@ export(ssrc_get)
 export(templeton_df)
 export(templeton_get)
 export(usaspend_get)
-importFrom(readr,read_csv)
 importFrom(rvest,"%>%")
