### 2021-06-04 PESData/awardFindR#25 @mccallc
I was personally more inclined to the autocomplete argument for `neh_get` rather than `get_neh` format, but it looks like the verb_subject format is more common in rOpenSci projects. Examples:

https://docs.ropensci.org/FedData/reference/index.html
https://docs.ropensci.org/c14bazAAR/reference/index.html

I'll change the function names to reflect this trend.

Also: changed the main function name from `awardFindR` to `award_search` as a trial balloon to see what bugs would pop up. If we're going for this format, though, I assume we want something more like `search_awards`?
