### 2021-06-04 b687ac4
testing longer http timeout for windows CI
