### 2021-06-04 PESData/awardFindR#25 @adam3smith
> I'll change the function names to reflect this trend.
Yeah, that makes sense then.

> Also: changed the main function name from awardFindR to award_search as a trial balloon to see what bugs would pop up. If we're going for this format, though, I assume we want something more like search_awards?

Yes, let's do verb_noun throughout then
