+++ tests/testthat/test-fedreporter.R
@@ -1,6 +1,6 @@
 test_that("Expected results from fedreporter", {
   suppressMessages(vcr::use_cassette("fedreporter", {
-    fed <- .fedreporter_standardize("interviews", "2012-01-01", "2020-01-01",
+    fed <- .standardize_fedreporter("interviews", "2012-01-01", "2020-01-01",
                                     FALSE)
   }))
   expect_equal(nrow(fed), 42)
