+++ R/rwjf.R
@@ -1,10 +1,10 @@
 #' Get Robert Wood Johnson foundation awards
-#' @inheritParams fedreporter_get
+#' @inheritParams get_fedreporter
 #' @return a data.frame
 #' @export
 #' @examples
-#' rwjf <- rwjf_get("qualitative data analysis", 2014, 2014)
-rwjf_get <- function(keyword, from_year, to_year, verbose=FALSE) {
+#' rwjf <- get_rwjf("qualitative data analysis", 2014, 2014)
+get_rwjf <- function(keyword, from_year, to_year, verbose=FALSE) {
   url <- paste0("https://www.rwjf.org/action/grants/database.json?",
   "k=", gsub(" ", "%20", keyword) , "&start=", from_year, "&end=", to_year,
   # Extra junk
@@ -56,7 +56,7 @@ rwjf_get <- function(keyword, from_year, to_year, verbose=FALSE) {
 }
 
 .rwjf_standardize <- function(keywords, from_date, to_date, verbose) {
-  raw <- lapply(keywords, rwjf_get,
+  raw <- lapply(keywords, get_rwjf,
                 format.Date(from_date, "%Y"), format.Date(to_date, "%Y"),
                 verbose)
   raw <- do.call(rbind.data.frame, raw)
