### 2021-06-14 eb5e3d9
correcting documentation with new function names
