+++ README.md
@@ -42,16 +42,16 @@ remotes::install_github("PESData/awardFindR")
 
 ## How the package works
 
-`award_search` has parameters to change keywords, sources and dates as search criteria, which are passed on to source routines. Dates are interpreted with varying degrees of precision based on the data available from each source. See included help on individual sources to understand their respective limitations.
+`search_awards` has parameters to change keywords, sources and dates as search criteria, which are passed on to source routines. Dates are interpreted with varying degrees of precision based on the data available from each source. See included help on individual sources to understand their respective limitations.
 
-Individual sources have functions of their own. `nsf_get` specifically fetches results from NSF, and `nih_get` fetches results from NIH, for example. These are meant to provide end users with higher levels of detail if they're interested in a specific source.
+Individual sources have functions of their own. `get_nsf` specifically fetches results from NSF, and `get_nih` fetches results from NIH, for example. These are meant to provide end users with higher levels of detail if they're interested in a specific source.
 
 ### Quick introduction
 
 Search for all awards matching a keyword since the default cutoff of Jan 1, 2019 to today
 
 ```
-awardFindR::award_search(keywords="ethnography")
+awardFindR::search_awards(keywords="ethnography")
 ```
 
 See the included vignette for additional examples.
@@ -59,7 +59,7 @@ See the included vignette for additional examples.
 For those interested in the results from a specific source, each source has its own function. For example, someone interested in NSF results for "ethnography" between 2018 and 2020 could run the following:
 
 ```
-awardFindR::nsf_get("ethnography", "2018-01-01", "2020-01-01")
+awardFindR::get_nsf("ethnography", "2018-01-01", "2020-01-01")
 ```
 
 Similar functions exist for each supported source. See included help for further details, as the arguments differ slightly between each.
