### 2021-06-14 #432 @mccallc
Hello!

Thank you very much @karawoo and @zambujo for your reviews and extensive comments. I really appreciate them. I've implemented changes that should address each individual comment, including:

Based on review from @karawoo:
- Adding a verbose option to all functions to print HTTP messages, defaulting to false (don't print the messages).
- Fixing documentation for functions like `gates_get` through the use of roxygen2 inheritParams
- implementing `match.args` for argument validation
- including output messages when there's no results from a source
- fixing all bugs that cause problems in the examples
- removing the global setting of `stringsAsFactors` entirely, which doesn't seem to cause any problems in oldrel versions of R.

Based on review from @zambujo:
- Using roxygen2 inheritParams to reduce repetition in documentation and harmonize between functions
- Moving installation instructions to the top of the readme
- Getting rid of read_csv dependency (removing readr dependency causes problems with `httr::content`, which needs it)
- changing function names to verb_object format. `awardFindR` is now `search_awards`, `nsf_get` is now `get_nsf`, etc

Again, thank you so all much. These comments really helped push this package in the right direction.

Let me know what's next in the process, @annakrystalli !
