### 2021-06-17 #432 @ropensci-review-bot
Approved! Thanks @mccallc for submitting and @karawoo, @zambujo for your reviews! :grin:
